requirements.dev.txt
absl-py==2.0.0
alabaster==0.7.13
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.0.0
appdirs==1.4.3
asks==2.3.3
async-exit-stack==1.0.1
async-generator==1.10
atomicwrites==1.4.1
attrs==19.1.0
Babel==2.13.0
beautifulsoup4==4.12.2
blinker==1.6.3
box2d-py==2.3.5
brax==0.9.2
bs4==0.0.1
bump2version==0.5.10
cachetools==5.3.1
carl-bench @ file:///home/rajanro/projects/rl-msc-pro/CARL
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
chex==0.1.83
click==8.1.7
cloudpickle==3.0.0
cmake==3.27.7
ConfigArgParse==1.7
ConfigSpace==0.7.1
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.36
-e git+https://github.com/Farama-Foundation/d4rl.git@71a9549f2091accff93eeff68f1f3ab2c0e0a288#egg=D4RL
dataclasses==0.6
DataProperty==1.0.1
dm-control==1.0.14
dm-env==1.6
dm-tree==0.1.8
docutils==0.14
etils==1.5.1
exceptiongroup==1.1.3
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.0
flax==0.7.4
fonttools==4.43.1
fsspec==2023.9.2
glfw==2.6.2
google-auth==2.23.3
google-auth-oauthlib==1.0.0
grpcio==1.59.0
gym==0.23.1
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
h5py==3.10.0
idna==3.4
imageio==2.31.5
imagesize==1.4.1
importlib-metadata==0.18
importlib-resources==6.1.0
install-jdk==1.1.0
itsdangerous==2.1.2
jax==0.4.18
jaxlib==0.4.18
jaxopt==0.8.1
jdk4py==17.0.7.0
Jinja2==3.1.2
kiwisolver==1.4.5
labmaze==1.0.6
lit==17.0.3
lxml==4.9.3
mario==0.0.150
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.0.0rc8
matplotlib==3.8.0
mbstrdecoder==1.1.3
mdurl==0.1.2
mjrl @ git+https://github.com/aravindr93/mjrl@3871d93763d3b49c4741e6daeaebbc605fe140dc
ml-collections==0.1.1
ml-dtypes==0.3.1
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
mujoco==2.3.7
mujoco-py==2.1.2.14
mypy==1.6.1
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.26.1
numpyencoder==0.3.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.1
outcome==1.0.0
packaging==23.2
pandas==2.1.1
parso==0.4.0
pathvalidate==3.2.0
Pillow==10.1.0
pip-autoremove==0.10.0
pluggy==0.13.1
protobuf==4.24.4
py==1.11.0
py4j==0.10.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybullet==3.2.5
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pygame==2.5.2
pyglet==2.0.9
Pygments==2.16.1
PyOpenGL==3.1.7
pyparsing==3.1.1
pyrsistent==0.15.2
pytablewriter==1.2.0
pytest==5.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytinyrenderer==0.0.14
pytz==2023.3.post1
PyVirtualDisplay==3.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
scipy==1.11.3
Shimmy==1.3.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stable-baselines3==2.1.0
swig==4.1.1
sympy==1.12
tabledata==1.3.3
tabulate==0.9.0
tcolorpy==0.1.4
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorboardX==2.6.2.2
tensorstore==0.1.46
termcolor==2.3.0
toml==0.10.0
tomli==2.0.1
toolz==0.9.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
trimesh==4.0.0
trio==0.11.0
trio-typing==0.2.0
triton==2.0.0
typepy==1.3.2
typing_extensions==4.8.0
tzdata==2023.3
UNKNOWN @ file:///home/rajanro/projects/rl-msc-pro
urllib3==2.0.7
wcwidth==0.2.8
Werkzeug==3.0.0
xmltodict==0.12.0
zipp==3.17.0

# Clone following packages manually for mario set up
# git clone https://github.com/frederikschubert/Mario-AI-Framework carl/envs/mario/Mario-AI-Framework
# git clone https://github.com/Mawiszus/TOAD-GUI  carl/envs/mario/TOAD-GUI

# System requirements
# sudo apt install libfreetype6-dev xvfb

# Compile java source files
# cd carl/envs/mario/Mario-AI-Framework/src
# javac *.java

# If CARL installed in UN_EDITABLE manner, then clone them into site packages 